Density map refactoring
